requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
Flask-Limiter==3.5.0